% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{brooksPars}
\alias{brooksPars}
\title{Model parameter values of Brooks et al. JTB 2012}
\format{A vector of parameter values.  \describe{
See description in comments in brooksPars.R of the doc folder. }}
\source{
The parameter values were taken from Table 1 of the PDF of the JTB 2012 paper.
}
\description{
This is a numeric vector that holds the parameter values of the model of Brooks et al. JTB 2012.
}
\note{
Coding supported by the National Cancer Institute and Tufts
Integrative Cancer Biology Program U54CA149233-029689.
}
\examples{

library(myelo)
brooksPars

}
\references{
Grace Brooks, Gabriel Provencher,  Jinzhi Lei, and Michael C. Mackey,
Neutrophil dynamics after chemotherapy and G-CSF: The role
of pharmacokinetics in shaping the response, \emph{Journal of Theoretical Biology} 
\bold{315} 97?109 (2012).
}
\seealso{
\code{\link{myelo-package}, \link{rcj12}, \link{fokas91},\link{scholz12},
\link{zhugePars},\link{raue10} }
}
\keyword{datasets}
