% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clapp15.R
\name{clapp15}
\alias{clapp15}
\title{CML-T-Cell Model of Clapp et al 2015}
\usage{
clapp15(Time, State, Pars)
}
\arguments{
\item{Time}{The parameters of this model do not depend on time: this argument
exists here as a dummy only because deSolve expects it to exist.}

\item{State}{Vector of current states. The elements of this vector must be
named because within this function's definition is
\code{with(as.list(State,Pars), {XXX} )} and code in XXX presupposes that it
can call named elements as variables with those names.}

\item{Pars}{Vector of parameter values with \emph{named} (see above) elements.}
}
\value{
A list of length 2 (as expected by deSolve) where the first list
 element is the vector of derivatives, i.e. the right hand side of the ODEs,
 and the second element of the list is a vector of auxiliary variables that
 one may wish to track over time.
}
\description{
This function returns the right hand side of an ordinary differential equation
model of T cells killing CML cells.  The intended use of this function is as
an argument to \code{ode()} of the \code{deSolve} package.
}
\references{
Geoffrey D. Clapp, Thomas Lepoutre, Raouf El Cheikh, Samuel Bernard, Jeremy Ruby, 
 Helene Labussie Wallet, Franck E. Nicolini, and Doron Levy, Implication of the Autologous
 Immune System in BCR-ABL Transcript Variations in Chronic Myelogenous
 Leukemia Patients Treated with Imatinib, \emph{Cancer Research},  \bold{75}, 4053-4062 (2015)
}
\seealso{
\code{\link{myelo-package}, \link{michor05}}
}
\author{
Tom Radivoyevitch
}
\keyword{IO}
