% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{claudianiPon}
\alias{claudianiPon}
\title{TFR Data (post-stopping ponatinib) in Figure 1 of S. Claudiani et al 2020}
\format{
Month is from Dx, Prct is BCR-ABL1 percent, Pt is patient ID, Ponatinib is its dose, T315I is its percent
}
\source{
Figure 1 (patients 4, 9 and 10).
}
\description{
This is a dataframe of BCR-ABL percentages for 3 CML patients stopping ponatinib.
}
\note{
Coding supported by the Cleveland Clinic Foundation.
}
\examples{

library(myelo)
head(claudianiPon)

}
\references{
Simone Claudiani et al. Prolonged treatment-free remission in chronic 
myeloid leukemia patients with previous BCR-ABL1 kinase domain mutations
\emph{Haematologica}, 2020 (\bold{105}): e227
}
\keyword{datasets}
