% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{colijnPars05a}
\alias{colijnPars05a}
\title{Model parameter values of Colijn and Mackey Periodic CML model (JTB 2005, part I)}
\format{A vector of parameter values.  \describe{
See description in comments in colijnPars05a.R of the doc folder. }}
\source{
The parameter values were taken from Table 1 of this  paper.
}
\description{
This is a numeric vector that holds the parameter values of Part I of the model of 
Colijn and Mackey in JTB 2005.
}
\note{
Coding supported by the Cleveland Clinic Foundation.
}
\examples{

library(myelo)
colijnPars05a

}
\references{
Caroline Colijn and Michael C. Mackey,
A mathematical model of hematopoiesis: I. Periodic chronic myelogenous leukemia, 
\emph{Journal of Theoretical Biology} 
\bold{237} 117-132 (2005).
}
\seealso{
\code{\link{colijnPars05b},\link{zhugePars},\link{brooksPars} }
}
\keyword{datasets}
