% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craig15.R
\name{craig15}
\alias{craig15}
\title{Neutrophil-GCSF-Chemo model of Craig et al 2015}
\usage{
craig15(Time, State, Pars)
}
\arguments{
\item{Time}{The current time.}

\item{State}{Vector of current states. The elements of this vector must be
named because within this function's definition is
\code{with(as.list(State,Pars), {XXX} )} and code in XXX presupposes that it
can call named elements as variables with those names.}

\item{Pars}{Vector of parameter values with \emph{named} (see above)
elements.}
}
\value{
A list of length 1 where the first list
element is the vector of derivatives, i.e. the right hand side of the ODEs
}
\description{
This function returns the right hand side of an ordinary differential delay
equation model of Craig et al. JTB 2015. 
The intended use of this function is as an argument to \code{dede()} of the \code{deSolve} package.
}
\details{
In this model Q = stem cells, Nr = marrow reserve neutrophils, and N = circulating neutrophils
}
\note{
This work was supported by the Cleveland Clinic Foundation.
}
\examples{

\dontrun{
library(deSolve)
library(myelo)
craigPars15
times <- seq(-30,200,by=0.1)
yout <- dede(c(Q=craigPars15[["Qss"]],Nr=craigPars15[["Nrss"]],N=craigPars15[["Nss"]]),
		times = times, func = craig15,	parms = craigPars15)
plot(yout)
}

}
\references{
# Morgan Craig ... Michael C. Mackey, Neutrophil dynamics during
 concurrent chemotherapy and G-CSF administration: Mathematical modelling
 guides dose optimisation to minimise neutropenia, \emph{Journal of
 Theoretical Biology} \bold{385} 77-89  (2015).
}
\seealso{
\code{\link{myelo-package}, \link{craig15}}
}
\author{
Tom Radivoyevitch (\email{txr24@case.edu})
}
\keyword{IO}
