% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craig16.R
\name{craig16}
\alias{craig16}
\title{Neutrophil-GCSF-Chemo model of Craig et al 2016}
\usage{
craig16(Time, State, Pars)
}
\arguments{
\item{Time}{The current time.}

\item{State}{Vector of current states. The elements of this vector must be
named because within this function's definition is
\code{with(as.list(State,Pars), {XXX} )} and code in XXX presupposes that it
can call named elements as variables with those names.}

\item{Pars}{Vector of parameter values with \emph{named} (see above)
elements.}
}
\value{
A list of length 1 where the first list
element is the vector of derivatives, i.e. the right hand side of the ODEs
}
\description{
This function returns the right hand side of an ordinary differential delay
equation model of Craig et al. Bull Math Biol 2016. 
The intended use of this function is as an argument to \code{dede()} of the \code{deSolve} package.
}
\details{
In this model Q = stem cells, Nr = marrow reserve neutrophils, and N = circulating neutrophils
}
\note{
This work was supported by the Cleveland Clinic Foundation.
}
\examples{

\dontrun{
library(deSolve)
library(myelo)
craigPars16
times <- seq(-30,200,by=.2)
yout <- dede(c(Q=craigPars16[["Qss"]],Nr=craigPars16[["Nrss"]],N=craigPars16[["Nss"]]),
		times = times, func = craig16,	parms = craigPars16)
plot(yout)
}

}
\references{
# M. Craig, A. R. Humphries, M. C. Mackey, A Mathematical Model of 
Granulopoiesis Incorporating the Negative Feedback Dynamics and Kinetics of 
G-CSF/Neutrophil Binding and Internalization, \emph{Bull Math Biol} \bold{78} 2304-2357  (2016).
}
\seealso{
\code{\link{myelo-package}, \link{craig15}}
}
\author{
Tom Radivoyevitch (\email{txr24@case.edu})
}
\keyword{IO}
