% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{craigPars15}
\alias{craigPars15}
\title{Model parameters  of Morgan Craig et al JTB 2015}
\format{A vector of parameter values.  \describe{ See description in comments
 in craigPars15.R of the doc folder. }}
\source{
The parameter values were taken from Table 1 of this  paper.
}
\description{
This is a numeric vector that holds the parameter values of the neutrophil
chemo GCSF model of Morgan Craig et al JTB 2015
}
\note{
Coding supported by the Cleveland Clinic Foundation.
}
\examples{

library(myelo)
craigPars15

}
\references{
# Morgan Craig ... Michael C. Mackey, Neutrophil dynamics during
 concurrent chemotherapy and G-CSF administration: Mathematical modelling
 guides dose optimisation to minimise neutropenia, \emph{Journal of
 Theoretical Biology} \bold{385} 77-89  (2015).
}
\seealso{
\code{\link{craigPars16},\link{zhugePars},\link{brooksPars} }
}
\keyword{datasets}
