% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{craigPars16}
\alias{craigPars16}
\title{Model parameters  of Morgan Craig et al BMB 2016}
\format{
A vector of parameter values.  \describe{ See description in comments
 in craigPars16.R of the doc folder. }
}
\source{
The parameter values were taken from Table 4-6 of this  paper.
}
\description{
This is a numeric vector that holds the parameter values of the neutrophil
chemo GCSF model of Morgan Craig et al BMB 2016
}
\note{
Coding supported by the Cleveland Clinic Foundation.
}
\examples{

library(myelo)
craigPars16

}
\references{
M. Craig, A. R. Humphries, M. C. Mackey, A Mathematical Model of 
Granulopoiesis Incorporating the Negative Feedback Dynamics and Kinetics of 
G-CSF/Neutrophil Binding and Internalization, \emph{Bull Math Biol} \bold{78} 2304-2357  (2016).
}
\seealso{
\code{\link{craigPars15},\link{zhugePars},\link{brooksPars} }
}
\keyword{datasets}
