% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{fribergPars02}
\alias{fribergPars02}
\title{Model parameters  of Friberg et al JCO 2002}
\format{
A vector of parameter values.  \describe{ See description in comments
 in fribergPars02.R of the doc folder. }
}
\source{
The parameter values were taken from the first row of Table 4 in the first paper and
for docetaxel PK, the first column of table II of the second paper.
}
\description{
This is a numeric vector that holds the parameter values of the neutrophil
chemo  model of Friberg et al JCO 2002
}
\note{
Coding supported by the Cleveland Clinic Foundation.
}
\examples{

library(myelo)
fribergPars02

}
\references{
Lena E. Friberg, Anja Henningsson, Hugo Maas, Laurent Nguyen, and Mats O. Karlsson
Model of Chemotherapy-Induced Myelosuppression With Parameter Consistency Across Drugs, 
\emph{Journal of Clinical Oncology},  \bold{20},  4713-4721 (2002)

Rene Bruno, Nicole Vivier, Jean Claude Vergniol,Susan L. De Phillips,
Guy Montay,  and Lewis B. Sheiner
A Population Pharmacokinetic Model for Docetaxel (Taxotere| Model Building and Validation 
\emph{Journal of Pharmacokinetics and Biopharmaceutics},  \bold{24},  153-172 (1996)
}
\seealso{
\code{\link{craigPars15},\link{zhugePars},\link{brooksPars} }
}
\keyword{datasets}
