% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{glauchePars20}
\alias{glauchePars20}
\title{Model parameters  of Tom Hahnel ... Ingmar Glauche, Bioarxiv 2020}
\format{
A dataframe of parameter values.  \describe{ See paper }
}
\source{
Supplement of Bioarxiv paper.
}
\description{
This is a dataframe of parameter values for 21 CML patients.
}
\note{
Coding supported by the Cleveland Clinic Foundation.
}
\examples{

library(myelo)
glauchePars20

}
\references{
Tom Hahnel, Christoph Baldow, Artur C. Fassoni , Joelle Guilhot , Francois Guilhot , Susanne
Inferring immunological control mechanisms from TKI dose alterations in CML patients, 
\emph{BioArxiv},  \bold{http://dx.doi.org/10.1101/722546},   (2020)
}
\keyword{datasets}
