% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{hahnelFigS5}
\alias{hahnelFigS5}
\title{TFR (post-stopping TKI) Data in Figure S5 of Hahnel et al}
\format{
Months is time from Dx, Prct is BCR-ABL1 percentage, Pt is patient ID, and UL is 1 if 
Prct is the upper limit of a left-censored measurement, else 0  \describe{ See paper }
}
\source{
Supplemental Figure S5.
}
\description{
This is a dataframe of BCR-ABL percentages for 21 CML patients.
}
\note{
Coding supported by the Cleveland Clinic Foundation.
}
\examples{

library(myelo)
head(hahnelFigS5)

}
\references{
Tom Hahnel et al. Model-Based Inference and Classification of 
Immunologic Control Mechanisms from TKI Cessation and Dose Reduction in Patients with CML
\emph{Cancer Research},  \bold{DOI: 10.1158/0008-5472.CAN-19-2175}, 2020;80:2394-406
}
\keyword{datasets}
