% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kim08.R
\name{kim08}
\alias{kim08}
\title{CML-T-Cell Model of Kim et al 2008}
\usage{
kim08(t, y, p)
}
\arguments{
\item{t}{The current time in the simulation.}

\item{y}{Vector of current states.}

\item{p}{Vector of parameter values with \emph{named} (see above) elements.}
}
\value{
A vector of derivatives (as expected by PBSddeSolve)
}
\description{
This function returns the right hand side of a delay differential equation
model of T cells killing CML cells.  The intended use of this function is as
an argument to \code{dde()} of the \code{PBSdeSolve} package.
}
\references{
Kim PS, Lee PP, Levy D, Dynamics and Potential Impact of the Immune
Response to Chronic Myelogenous Leukemia, \emph{PLoS Comput Biol}, \bold{4}, 6 (2008)
}
\seealso{
\code{\link{myelo-package}, \link{clapp15}, \link{michor05}}
}
\author{
Tom Radivoyevitch and Rainer Sachs
}
\keyword{IO}
