% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{michor}
\alias{michor}
\title{Michor's Nature 2005 Data}
\format{
Prct is BCR-ABL1 percent, Pt is patient ID, and Days is presumably from first Tx, not Dx.
}
\source{
Figure 5 (in the supplement).
}
\description{
This is a dataframe of BCR-ABL percentages for 68 CML patients taking imatinib up to one year.
}
\examples{

library(myelo)
head(michor)

}
\references{
Franziska Michor, Timothy P. Hughes, Yoh Iwasa, Susan Branford,
Neil P. Shah, Charles L. Sawyers & Martin A. Nowak
Dynamics of chronic myeloid leukaemia, \emph{Nature},
\bold{435}, 1267-1270 (2005)
}
\keyword{datasets}
