% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{miSR}
\alias{miSR}
\title{Michor's Stop and Relapse 2005 Data}
\format{
Prct is BCR-ABL1 percent, Pt is patient ID, and Days is presumably from first Tx, not Dx, 
DOSE of TKI was replaced with a Ponatinib equivalent of 45 mg. MDV (missing dependent variable) is 
1 (TRUE) at TKI stopping times at which
BCR-ABL was not also measured (this happens in the first 2 of the 3 pts that stopped TKI);
}
\source{
Figures 2 & 3.
}
\description{
This is a dataframe of BCR-ABL percentages of 3 CML patients stopping TKI in Figure 2 and 4 
patients relapsing in Figure 3.
}
\examples{

library(myelo)
head(miSR)

}
\references{
Franziska Michor, Timothy P. Hughes, Yoh Iwasa, Susan Branford,
Neil P. Shah, Charles L. Sawyers & Martin A. Nowak
Dynamics of chronic myeloid leukaemia, \emph{Nature},
\bold{435}, 1267-1270 (2005)
}
\keyword{datasets}
