% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmar17.R
\name{parmar17}
\alias{parmar17}
\title{iron metabolism model of Parmar et al. 2017}
\usage{
parmar17(Time, State, Pars)
}
\arguments{
\item{Time}{The parameters of this model do not depend on time: this argument
exists here as a dummy only because deSolve expects it to exist.}

\item{State}{Vector of current states. The elements of this vector must be
named because within this function's definition is
\code{with(as.list(State,Pars), {XXX} )} and code in XXX presupposes that it
can call named elements as variables with those names.}

\item{Pars}{Vector of parameter values with \emph{named} (see above) elements.}
}
\value{
A list of length 2 (as expected by deSolve) where the first list
 element is the vector of derivatives, i.e. the right hand side of the ODEs,
 and the second element of the list is a vector of auxiliary variables that
 one may wish to track over time.
}
\description{
This function returns the right hand side of an ordinary differential equation
model of iron metabolism in mouse fed a normal iron diet, published by Parmar
et al in BMC Systems Biology in 2017.  The intended use of this function is as
an argument to \code{ode()} of the \code{deSolve} package.
}
\references{
Jignesh H. Parmar, Grey Davis, Hope Shevchuk and Pedro Mendes,
 Modeling the dynamics of mouse iron body distribution: hepcidin is necessary
 but not sufficient,  \emph{BMC Systems Biology}, \bold{11}, 57 (2017)
}
\seealso{
\code{\link{myelo-package}, \link{fokas91}}
}
\author{
Tom Radivoyevitch
}
\keyword{IO}
