% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmar19.R
\name{parmar19}
\alias{parmar19}
\title{iron metabolism model of Parmar and Mendes 2019}
\usage{
parmar19(Time, State, Pars)
}
\arguments{
\item{Time}{The parameters of this model do not depend on time: this argument
exists here as a dummy only because deSolve expects it to exist.}

\item{State}{Vector of current states. The elements of this vector must be
named because within this function's definition is
\code{with(as.list(State,Pars), {XXX} )} and code in XXX presupposes that it
can call named elements as variables with those names.}

\item{Pars}{Vector of parameter values with \emph{named} (see above) elements.}
}
\value{
A list of length 2 (as expected by deSolve) where the first list
 element is the vector of derivatives, i.e. the right hand side of the ODEs,
 and the second element of the list is a vector of auxiliary variables that
 one may wish to track over time.
}
\description{
This function returns the right hand side of an ordinary differential equation
model of iron metabolism in mice published by Parmar and Mendes
in Plos Comp Biol in 2019.  The intended use of this function is as
an argument to \code{ode()} of the \code{deSolve} package. To emulate hemachromatosis set ksHepci to 0.
Our goal here is to reproduce some of the Table 1 values at 365 days.
}
\references{
Jignesh H. Parmar and Pedro Mendes,
 A computational model to understand mouse iron physiology and disease,  \
 emph{PLoS Comp Biol}, \bold{15}, 1 (2019)
}
\seealso{
\code{\link{myelo-package}, \link{parmar17}}
}
\author{
Tom Radivoyevitch
}
\keyword{IO}
