\docType{data}
\name{parmarPars}
\alias{parmarPars}
\title{Model parameter values of parmarPars et al. BMCSB 2017}
\format{A vector of parameter values.  \describe{
See description in comments in parmarPars.R of the doc folder. }}
\source{
Jignesh H. Parmar, Grey Davis, Hope Shevchuk and Pedro Mendes,
Modeling the dynamics of mouse iron body distribution: hepcidin is necessary
but not sufficient,  BMC Systems Biology, 11:57 (2017)
these estimates all correspond to column #6 in Table 2
}
\description{
  This is a numeric vector that holds the parameter values
  of the model of Parmar et al. BMC Systems Biology 2017.
}
\examples{
library(myelo)
parmarPars
}
\references{

Jignesh H. Parmar, Grey Davis, Hope Shevchuk and Pedro Mendes,
Modeling the dynamics of mouse iron body distribution: hepcidin is necessary
but not sufficient,  \emph{BMC Systems Biology}, \bold{11}, 57 (2017)
}
\seealso{
  \code{\link{myelo-package}, \link{rcj12},
  \link{fokas91},\link{scholz12},
  \link{zhugePars},\link{raue10} }
}
\keyword{datasets}

