\docType{data}
\name{parmarPars19}
\alias{parmarPars19}
\title{Model parameter values of Parmar and Mendes Plos Comp Biol 2019}
\format{A vector of parameter values.  \describe{
See description in comments in parmarPars19.R of the doc folder. }}
\source{
Jignesh H. Parmar and Pedro Mendes,
A computational model to understand mouse iron physiology and disease,
emph{PLoS Comp Biol}, \bold{15}, 1 (2019) 
}
\description{
  This is a numeric vector that holds the parameter values
  of the model of Parmar and Mendes Plos Comp Biol 2019.
}
\examples{
library(myelo)
parmarPars19
}
\references{
Jignesh H. Parmar and Pedro Mendes,
A computational model to understand mouse iron physiology and disease,
emph{PLoS Comp Biol}, \bold{15}, 1 (2019) 
}
\seealso{
  \code{\link{myelo-package}, \link{parmar17},
  \link{parmar19} }
}
\keyword{datasets}

