% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{scholzPars}
\alias{scholzPars}
\title{Model parameter values of Scholz et al. TBMM 2012}
\format{A vector of parameter values.  \describe{
See description in comments in sholzPars.R of the doc folder. }}
\source{
The parameter values were taken from the supplement PDF of the TBMM 2012 paper.
}
\description{
This is a numeric vector of 128 parameter values of the model of Scholz et al. TBMM 2012.
}
\note{
Coding supported by the National Cancer Institute and Tufts
Integrative Cancer Biology Program U54CA149233-029689.
}
\examples{

library(myelo)
scholzPars

}
\references{
M. Scholz, S. Schirm, M. Wetzler, C. Engel
and M. Loeffler, Pharmacokinetic and -dynamic modelling of
G-CSF derivatives in humans, \emph{Theoretical Biology and Medical Modelling} 
\bold{9} 32 (2012).
}
\seealso{
\code{\link{myelo-package}, \link{rcj12}, \link{fokas91},\link{scholz12},
\link{raue10} }
}
\keyword{datasets}
