% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhuge19.R
\name{zhuge19}
\alias{zhuge19}
\title{Neutrophil and Platelet model of Zhuge et al 2019}
\usage{
zhuge19(Time, State, Pars)
}
\arguments{
\item{Time}{The current time.}

\item{State}{Vector of current states. The elements of this vector must be
named because within this function's definition is
\code{with(as.list(State,Pars), {XXX} )} and code in XXX presupposes that it
can call named elements as variables with those names.}

\item{Pars}{Vector of parameter values with \emph{named} (see above)
elements.}
}
\value{
A list of length 1 where the first list
element is the vector of derivatives, i.e. the right hand side of the ODEs
}
\description{
This function returns the right hand side of an ordinary differential delay
equation model of Zhuge et al. JTB 2019. 
The intended use of this function is as an argument to \code{dede()} of the \code{deSolve} package.
}
\details{
In this model S = stem cells, N = neutrophils, and P = platlets
}
\note{
This work was supported by the Cleveland Clinic Foundation.
}
\examples{

\dontrun{
library(deSolve)
library(myelo)
zhugePars19
times <- seq(-(zhugePars19["tauPM"]+zhugePars19["tauPS"]),200,by=0.1)
yout <- dede(c(S=zhugePars19[["Sss"]],N=zhugePars19[["Nss"]],P=zhugePars19[["Pss"]]),
		times = times, func = zhuge19,	parms = zhugePars19)
plot(yout)
}

}
\references{
Changjing Zhuge, Michael C. Mackey, and Jinzhi Lei,
Origins of oscillation patterns in cyclical thrombocytopenia, 
\emph{Journal of Theoretical Biology} 
\bold{462} 432-445 (2019).
}
\seealso{
\code{\link{myelo-package}, \link{zhuge12}}
}
\author{
Tom Radivoyevitch (\email{txr24@case.edu})
}
\keyword{IO}
