\docType{data}
\name{zhugePars}
\alias{zhugePars}
\title{Model parameter values of Zhuge et al. JTB 2012}
\format{A vector of parameter values.  \describe{
See description in comments in zhugePars.R of the doc folder. }}
\source{
  The parameter values were taken from Table 1 of the PDF
  of the JTB 2012 paper.
}
\description{
  This is a numeric vector that holds the parameter values
  of the model of Zhuge et al. JTB 2012.
}
\note{
  Coding supported by the National Cancer Institute and
  Tufts Integrative Cancer Biology Program
  U54CA149233-029689.
}
\examples{
library(myelo)
zhugePars
}
\references{
  Changjing Zhuge, Jinzhi Lei, and Michael C. Mackey,
  Neutrophil dynamics in response to chemotherapy and
  G-CSF, \emph{Journal of Theoretical Biology} \bold{293}
  111?120 (2012).
}
\seealso{
  \code{\link{myelo-package}, \link{rcj12},
  \link{fokas91},\link{scholz12},
  \link{brooksPars},\link{raue10} }
}
\keyword{datasets}

