% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{data}
\name{zhugePars19}
\alias{zhugePars19}
\title{Model parameter values of Zhuge et al. JTB 2019}
\format{
A vector of parameter values.  \describe{
See description in comments in zhugePars19.R of the doc folder. }
}
\source{
The parameter values were taken from Table 1 of this 2019 paper.
}
\description{
This is a numeric vector that holds the parameter values of the model of Zhuge et al. JTB 2019.
}
\note{
Coding supported by the Cleveland Clinic Foundation.
}
\examples{

library(myelo)
zhugePars19

}
\references{
Changjing Zhuge, Michael C. Mackey, and Jinzhi Lei,
Origins of oscillation patterns in cyclical thrombocytopenia, \emph{Journal of Theoretical Biology} 
\bold{462} 432-445 (2019).
}
\seealso{
\code{\link{zhugePars},\link{brooksPars} }
}
\keyword{datasets}
