% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myelo-package.R
\docType{package}
\name{myelo-package}
\alias{myelo-package}
\title{Models of myeloid hematopoiesis}
\description{
This package aims to make analyses of models of myeloid hematopoiesis more 
readily reproducible. In the case of ODE models it provides functions that
compute the right hand sides of systems of ODEs.
}
\details{
\tabular{ll}{ Package: \tab \pkg{myelo}\cr Type: \tab Package\cr Depends:
\tab deSolve\cr Suggests: \tab bbmle \cr License: \tab GPL-2\cr LazyLoad:
\tab yes\cr LazyData: \tab yes\cr URL: \tab
\url{http://epbi-radivot.cwru.edu/myelo/myelo.html}\cr }
}
\note{
This work was supported by the National Cancer Institute and Tufts
Integrative Cancer Biology Program under U54CA149233-029689.
}
\seealso{
\code{\link{fokas91},\link{rcj12}}
}
\author{
Tom Radivoyevitch (\email{txr24@case.edu})
}
\keyword{package}
